requirements.txt
# ==========================================
# NETS Business Data Enhancement System
# Data Processing Pipeline + ML Models + Visualization
# Python 3.10+ (Anaconda environment)
# ==========================================

# === Core Data Processing ===
pandas==3.0.0                    # DataFrame operations + CSV handling
numpy==2.4.1                     # Numerical computing + arrays
geopandas==0.14.3                # Spatial data frames (EPSG:4326 operations)
shapely==2.0.4                   # Geometric operations (points, polygons)

# === Machine Learning + Statistical ===
scikit-learn==1.5.2              # Preprocessing, feature scaling, metrics
xgboost==2.1.3                   # XGBoost regression (employee estimation)
pymc==5.14.0                     # Bayesian hierarchical modeling + confidence intervals
arviz==0.19.0                    # ArviZ diagnostics for PyMC

# === API Clients ===
openai==2.15.0                   # GPT-4o-mini for review analysis
googlemaps==4.10.0               # Google Maps Places API
requests==2.32.5                 # HTTP requests
waybackpy==3.0.6                 # Wayback Machine API wrapper
beautifulsoup4==4.12.3           # HTML parsing

# === Enhanced Data Collection ===
outscraper==1.2.0                # Outscraper Google Maps reviews (unlimited)
linkedin-scraper>=3.1.0          # LinkedIn company data scraping
playwright>=1.40.0               # Browser automation (LinkedIn)

# === Text Processing + Matching ===
transformers==4.41.2             # Hugging Face models (zero-shot classification)
fuzzywuzzy==0.18.0               # Fuzzy string matching (entity deduplication)
python-Levenshtein==0.25.1       # Levenshtein distance (fuzzywuzzy backend)
usaddress==0.5.10                # US address parsing + normalization

# === Visualization + Dashboard ===
streamlit==1.40.2                # Interactive dashboard framework
folium==0.15.1                   # Leaflet maps (heatmaps, markers)
altair==5.4.1                    # Declarative visualization (time series)
matplotlib==3.9.2                # Static plots
seaborn==0.13.2                  # Statistical visualization

# === Configuration & Environment ===
python-dotenv==1.2.1             # Environment variable management
pydantic==2.12.5                 # Data validation + settings
pydantic-core==2.41.5            # Pydantic backend

# === Big Data Processing (optional, for >1000 records) ===
dask==2024.12.0                  # Distributed computing
dask-dataframe==2024.12.0        # Dask DataFrame

# === Computer Vision ===
opencv-python==4.11.0.86         # Image processing (street view facade analysis)

# === Geospatial ===
haversine==2.7.0                 # Great-circle distance calculation
geopy==2.4.1                     # Geocoding + distance utilities

# === File Format Support ===
pyarrow==17.0.0                  # Parquet serialization (output format)
openpyxl==3.1.5                  # Excel file support (optional reports)

# === HTTP & Network ===
httpx==0.28.1                    # Modern async HTTP client
httpcore==1.0.9                  # Low-level HTTP
h11==0.16.0                      # HTTP protocol parser

# === Utilities ===
tqdm==4.67.1                     # Progress bars for loops
python-dateutil==2.9.0.post0     # Date manipulation
pytz==2025.1                     # Timezone handling

# === Logging & Monitoring ===
loguru==0.7.2                    # Advanced logging (optional, better than standard logging)

# === Testing & Development ===
pytest==8.3.4                    # Unit testing framework
pytest-cov==6.0.0                # Code coverage
black==24.10.0                   # Code formatter
flake8==7.1.1                    # Linter

# === Dependencies (auto-installed) ===
certifi==2026.1.4                # SSL certificates
charset-normalizer==3.4.4        # Character encoding
idna==3.11                       # Internationalized domain names
urllib3==2.6.3                   # URL handling
sniffio==1.3.1                   # Async library detection
anyio==4.12.1                    # Async compatibility layer
jiter==0.12.0                    # JSON parsing
typing_extensions==4.15.0        # Type hint extensions
annotated_types==0.7.0           # Type annotations
pytz==2025.1                     # Timezone database
tzdata==2025.3                   # Timezone data (Windows)
six==1.17.0                      # Python 2/3 compatibility
colorama==0.4.6                  # Terminal color support
distro==1.9.0                    # OS detection
